requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.1
paramiko==3.3.1
Pint==0.22
pluggy==1.3.0
pycparser==2.21
pyfrc==2023.0.1
PyNaCl==1.5.0
pynetconsole==2.0.4
pyntcore==2023.4.3.0
pytest==7.4.0
pytest-reraise==2.1.2
robotpy==2023.4.3.0
robotpy-commands-v2==2023.4.3.0
robotpy-hal==2023.4.3.0
robotpy-halsim-gui==2023.4.3.0
robotpy-installer==2023.0.3
robotpy-rev==2023.1.3.2
robotpy-wpilib-utilities==2023.1.0
robotpy-wpimath==2023.4.3.0
robotpy-wpinet==2023.4.3.0
robotpy-wpiutil==2023.4.3.0
tomli==2.0.1
typing_extensions==4.7.1
wpilib==2023.4.3.0
